issue #281: implement solution with pinned acl cache to resolve problem with slow security checks
