+++ server/catgenome/profiles/dev/catgenome.properties
@@ -84,6 +84,7 @@ saml.security.enable=false
 #saml.authn.request.binding=urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Redirect
 #saml.authorities.attribute.names=http://schemas.xmlsoap.org/ws/2005/05/identity/claims/tokenGroups
 #saml.user.attributes=Email=http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress,Name=http://schemas.xmlsoap.org/ws/2005/05/identity/claims/name
+#security.acl.cache.period=86400
 
 # Create a NGB user if it is not present in the database. Available strategies: AUTO, EXPLICIT, EXPLICIT_GROUP
 saml.user.auto.create=EXPLICIT
