+++ server/catgenome/src/main/java/com/epam/catgenome/app/AclSecurityConfiguration.java
@@ -29,13 +29,10 @@ import static com.epam.catgenome.entity.user.DefaultRoles.*;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import javax.sql.DataSource;
 
 import com.epam.catgenome.security.acl.customexpression.NGBMethodSecurityExpressionHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.cache.ehcache.EhCacheFactoryBean;
-import org.springframework.cache.ehcache.EhCacheManagerFactoryBean;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.*;
 import org.springframework.security.access.PermissionEvaluator;
@@ -45,13 +42,9 @@ import org.springframework.security.access.hierarchicalroles.RoleHierarchyImpl;
 import org.springframework.security.acls.AclPermissionEvaluator;
 import org.springframework.security.acls.domain.*;
 import org.springframework.security.acls.jdbc.JdbcMutableAclService;
-import org.springframework.security.acls.jdbc.LookupStrategy;
-import org.springframework.security.acls.model.AclCache;
-import org.springframework.security.acls.model.PermissionGrantingStrategy;
 import org.springframework.security.acls.model.SidRetrievalStrategy;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.method.configuration.GlobalMethodSecurityConfiguration;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
 import com.epam.catgenome.entity.user.DefaultRoles;
 import com.epam.catgenome.security.acl.*;
@@ -61,12 +54,11 @@ import com.epam.catgenome.security.acl.*;
 @EnableGlobalMethodSecurity(securedEnabled = true, prePostEnabled = true)
 @ComponentScan(basePackages = "com.epam.catgenome.security.acl")
 @ImportResource("classpath*:conf/catgenome/acl-dao.xml")
+@Import(AclCacheConfiguration.class)
 public class AclSecurityConfiguration extends GlobalMethodSecurityConfiguration {
-    @Autowired
-    private ApplicationContext context;
 
     @Autowired
-    private DataSource dataSource;
+    private ApplicationContext context;
 
     @Autowired
     private PermissionFactory permissionFactory;
@@ -119,45 +111,4 @@ public class AclSecurityConfiguration extends GlobalMethodSecurityConfiguration 
         return new JdbcMutableAclServiceImpl(dataSource, lookupStrategy(), aclCache());
     }*/
 
-    @Bean
-    public LookupStrategy lookupStrategy() {
-        return new LookupStrategyImpl(dataSource, aclCache(), aclAuthorizationStrategy(),
-                                      auditLogger(), permissionFactory, permissionGrantingStrategy());
-    }
-
-    @Bean
-    public AuditLogger auditLogger() {
-        return new ConsoleAuditLogger();
-    }
-
-    @Bean
-    public AclAuthorizationStrategy aclAuthorizationStrategy() {
-        return new AclAuthorizationStrategyImpl(new SimpleGrantedAuthority(ROLE_ADMIN.getName()));
-    }
-
-    @Bean
-    public PermissionGrantingStrategy permissionGrantingStrategy() {
-        return new PermissionGrantingStrategyImpl(auditLogger());
-    }
-
-    @Bean
-    public AclCache aclCache() {
-        return new EhCacheBasedAclCache(ehCacheFactoryBean().getObject(),
-                permissionGrantingStrategy(), aclAuthorizationStrategy());
-    }
-
-    @Bean
-    public EhCacheFactoryBean ehCacheFactoryBean() {
-        EhCacheFactoryBean factoryBean = new EhCacheFactoryBean();
-        factoryBean.setCacheManager(ehCacheManagerFactoryBean().getObject());
-        factoryBean.setCacheName("aclCache");
-        return factoryBean;
-    }
-
-    @Bean
-    public EhCacheManagerFactoryBean ehCacheManagerFactoryBean() {
-        EhCacheManagerFactoryBean factoryBean = new EhCacheManagerFactoryBean();
-        factoryBean.setCacheManagerName("aclCacheManager");
-        return factoryBean;
-    }
 }
